+++ HISTORY.rst
@@ -4,14 +4,13 @@ History
 0.7.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- report event selection: if user is in invited list, add the event only if she/he accepted or not answered
 
 
 0.7.0 (2024-04-09)
 ------------------
 
 - Added ``read`` option to ``--execute``
-
 
 0.6.0 (2023-03-31)
 ------------------